refactor: convert block and pin APIs to async/await
