+++ src/core/components/index.js
@@ -3,12 +3,16 @@
 exports.add = require('./add')
 exports.config = require('./config')
 exports.init = require('./init')
+exports.pin = {
+  add: require('./pin/add'),
+  ls: require('./pin/ls'),
+  rm: require('./pin/rm')
+}
 exports.start = require('./start')
 exports.stop = require('./stop')
 
 exports.legacy = { // TODO: these will be removed as the new API is completed
   dag: require('./dag'),
   libp2p: require('./libp2p'),
-  object: require('./object'),
-  pin: require('./pin')
+  object: require('./object')
 }
